re-structuring my test files and adding hooks
