+++ server/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "babel src -d dist",
     "start": "npm run build && nodemon dist/index.js",
-    "test": "mocha src/test  --exit",
+    "test": "mocha src/test --timeout 50000 --exit",
     "coverage": "nyc npm run test",
     "make-badges": "istanbul-badges-readme --readmeDir='../'",
     "codecov": "nyc report --reporter=text-lcov > coverage.lcov && codecov"
